Refactor Duke Package name
