+++ src/main/java/Duke/Command/Event.java
@@ -1,4 +1,4 @@
-package Task;
+package Duke.Command;
 
 import java.text.SimpleDateFormat;
 import java.time.LocalDate;
@@ -6,7 +6,7 @@ import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 import java.util.Date;
 
-import TaskPackage.Tasks;
+import Duke.Task.Tasks;
 
 public class Event extends Tasks {
